requirements.txt
asgiref==3.8.1
click==8.1.7
cmake==3.31.0.1
colorama==0.4.6
Django==5.1.1
dlib==19.24.6
et_xmlfile==2.0.0
face-recognition==1.3.0
face_recognition_models==0.3.0
numpy==2.1.3
opencv-python==4.10.0.84
openpyxl==3.1.5
pillow==10.4.0
psycopg2==2.9.9
setuptools==75.3.0
sqlparse==0.5.1
tzdata==2024.2
wheel==0.45.0
